Adicionei a função de adicionar ou visualizar os endereço na parte dos administradores

+++ app/src/main/java/com/application/coletorplus/data/dao/ValidadeDao.java
@@ -23,7 +23,7 @@ public interface ValidadeDao {
     List<Validade> listarPorVencimento();
 
     @Query("SELECT * FROM validades WHERE produtoId = :produtoId ORDER BY dataVencimento ASC")
-    List<Validade> buscarPorProduto(int produtoId);
+    List<Validade> buscarPorProduto(long produtoId);
 
     @Delete
     void deletar(Validade validade);
